// src/types/element-types.ts
export enum ElementType {
  // Layout Elements
  LAYOUT = 'layout',
  SECTION = 'section',
  ROW = 'row',
  COLUMN = 'column',
  
  // Basic Elements
  HEADING = 'heading',
  TEXT = 'text',
  BUTTON = 'button',
  IMAGE = 'image',
  VIDEO = 'video',
  SPACER = 'spacer',
  DIVIDER = 'divider',
  COUNTDOWN = 'countdown',
  QRCODE = 'qrcode',
  ICON = 'icon',
  
  // Shape Elements
  SHAPE = 'shape',
  LINE = 'line',
  STAR = 'star',
  HEART = 'heart',
  CHECK = 'check',
  CROSS = 'cross',
  PLUS = 'plus',
  MINUS = 'minus',
  ARROW = 'arrow',
  PLAY = 'play',
  PAUSE = 'pause',
  STOP = 'stop',
  VOLUME = 'volume',
  MUTE = 'mute',
  SETTINGS = 'settings',
  USER = 'user',
  USERS = 'users',
  HOME = 'home',
  SEARCH = 'search',
  CALENDAR = 'calendar',
  CLOCK = 'clock',
  LOCATION = 'location',
  PHONE = 'phone',
  EMAIL = 'email',
  LINK = 'link',
  DOWNLOAD = 'download',
  UPLOAD = 'upload',
  EDIT = 'edit',
  DELETE = 'delete',
  SAVE = 'save',
  PRINT = 'print',
  SHARE = 'share',
  LIKE = 'like',
  DISLIKE = 'dislike',
  COMMENT = 'comment',
  REPLY = 'reply',
  FORWARD = 'forward',
  BACKWARD = 'backward',
  REFRESH = 'refresh',
  RELOAD = 'reload',
  CLOSE = 'close',
  OPEN = 'open',
  NEW = 'new',
  OLD = 'old',
  FIRST = 'first',
  LAST = 'last',
  NEXT = 'next',
  PREVIOUS = 'previous',
  UP = 'up',
  DOWN = 'down',
  LEFT = 'left',
  RIGHT = 'right',
  CENTER = 'center',
  TOP = 'top',
  BOTTOM = 'bottom',
  MIDDLE = 'middle',
  START = 'start',
  END = 'end',
  BEGIN = 'begin',
  FINISH = 'finish',
  COMPLETE = 'complete',
  INCOMPLETE = 'incomplete',
  SUCCESS = 'success',
  FAILURE = 'failure',
  ERROR = 'error',
  WARNING = 'warning',
  INFO = 'info',
  HELP = 'help',
  QUESTION = 'question',
  ANSWER = 'answer',
  YES = 'yes',
  NO = 'no',
  TRUE = 'true',
  FALSE = 'false',
  ON = 'on',
  OFF = 'off',
  ENABLE = 'enable',
  DISABLE = 'disable',
  ACTIVE = 'active',
  INACTIVE = 'inactive',
  VISIBLE = 'visible',
  HIDDEN = 'hidden',
  SHOW = 'show',
  HIDE = 'hide',
  DISPLAY = 'display',
  NONE = 'none',
  BLOCK = 'block',
  INLINE = 'inline',
  INLINE_BLOCK = 'inline-block',
  FLEX = 'flex',
  GRID = 'grid',
  TABLE = 'table',
  LIST_ITEM = 'list-item',
  RUN_IN = 'run-in',
  RUN_OUT = 'run-out',
  FLOW = 'flow',
  FLOW_ROOT = 'flow-root',
  CONTENTS = 'contents',
  INITIAL = 'initial',
  INHERIT = 'inherit',
  UNSET = 'unset',
  REVERT = 'revert',
  REVERT_LAYER = 'revert-layer'
}